Fix merging of documents with attachment stubs
During replicated updates, merging of documents with
attachment stubs will fail if, after merging the received
document's revision tree with the current on disk revision
tree, produces a revision tree which doesn't contain the revision
that immediately precedes the received document's revision.

This happens when the received document doesn't contain in its
revision history any of the revisions in the revision tree
of the currently on disk document. This is possible when the
document had a number of updates higher than the value of revs
limit defined for the target database.

Fixed by Filipe David Borba Manana
See https://issues.apache.org/jira/browse/COUCHDB-1365